promote TracedRuntime instead of TelemetryRuntime
