+++ dial9-tokio-telemetry/src/telemetry/mod.rs
@@ -24,8 +24,8 @@ pub use format::{
 };
 pub use recorder::{
     HasTracePath, NoTracePath, RuntimeTelemetryHandle, TelemetryCore, TelemetryGuard,
-    TelemetryHandle, TraceRuntimeCoreBuilder, TracedRuntime, TracedRuntimeBuilder,
-    current_worker_id,
+    TelemetryHandle, TelemetryRuntimeError, TraceRuntimeCoreBuilder, TracedRuntime,
+    TracedRuntimeBuilder, current_worker_id,
 };
 pub use task_metadata::{TaskId, UNKNOWN_TASK_ID};
 pub use writer::{NullWriter, RotatingWriter, TraceWriter};
